cloud-function/requirements.txt
google-cloud-firestore
google-cloud-storage
PyMuPDF==1.23.6